+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # 1.0.1.9000
 
+* `consume()` now has a `poll_interval` argument to set how often to poll
+  the queue for new jobs.
+
 # 1.0.1
 
 * Set the `LITEQ_CACHE_DIR` environment variable to change the